Made things a bit slicker by using only a single routeController for the traits since they all share pretty much all characteristic together.
It is still modular enough to support own data-rows etc. (Or will be, the save function has to be extended first)

Most of the time was spend fighting with xenforo so ultimately there isn't even that much code. Literally.

Also made a interface for the traits that will have common methods they share with each other.

The two documents mastery_trait_edit and mastery_traits basically are taken by example from the routing controller.

If there are questions lemme know.